% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genNA.R
\name{genNAs}
\alias{genNAs}
\title{Añade valores faltantes a un vectir, matriz o data frame}
\usage{
genNAs(x, p = 0.1, seed = NULL)
}
\arguments{
\item{x}{Un vector, matriz o \code{data.frame}.}

\item{p}{Proporcion de valores faltantes para añadir al \code{x}. En caso \code{x} es un \code{data.frame}, \code{p} puede tambien ser un vector de probabilidades por columna o un vector nombrado (ver ejemplos).}

\item{seed}{Planta una semilla.}
}
\value{
\code{x} con valores faltantes.
}
\description{
Crea o importa un vector, matriz o data frame y reemplaza algunos de sus valores por NAs.
}
\examples{
genNAs(1:10, p = 0.5, seed = 3345)
genNAs(rep(Sys.Date(), 10))
genNAs(cbind(1:10, 10:1), p = 0.2)
head(genNAs(iris))
head(genNAs(iris, p = 0.2))
head(genNAs(iris, p = c(0, 1, 0.5, 0.5, 0.5)))
head(genNAs(iris, p = c(Sepal.Length = 1)))
head(genNAs(iris, p = c(Species = 0.2, Sepal.Length = 0.5)))
}
\author{
Julian Flores y Leonela Yamuca
}
