% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impUniv.R
\name{impUniv}
\alias{impUniv}
\title{Imputacion Univariada}
\usage{
impUniv(x, v = NULL, seed = NULL)
}
\arguments{
\item{x}{Un vector, matriz o data frame.}

\item{v}{Un vector caracter de nombre de columna a imputar (solo relevante si \code{x} es un data frame). El defecto \code{NULL} imputa todas las columnas.}

\item{seed}{Una semilla.}
}
\value{
\code{x} con valores imputados.
}
\description{
Llena los valores faltantes de un vector, una matriz o data frame
por muestreo con reemplazamiento de los valores no faltantes. Para data frames,
el muestreo es propio de cada columna.
}
\examples{
impUniv(c(NA, 0, 1, 0, 1))
impUniv(c("A", "A", NA))
impUniv(as.factor(c("A", "A", NA)))
head(impUniv(genNAs(iris)))
head(impUniv(genNAs(iris), v = "Species"))
head(impUniv(genNAs(iris), v = c("Species", "Petal.Length")))
}
\author{
Julian Flores y Leonela Yamuca
}
