% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm.R
\name{pmm}
\alias{pmm}
\title{missRanger pmm}
\usage{
pmm(xtrain, xtest, ytrain, k = 1L, seed = NULL)
}
\arguments{
\item{xtrain}{Vector con valores predichos in la data de training. Puede ser de tipo logical, numeric, character, o factor.}

\item{xtest}{Vector como \code{xtrain} con valores predichos en el test data. Valores faltantes no son permitidos.}

\item{ytrain}{Vector de los valores observados en el training data. Deben ser de la misma longitud que \code{xtrain}. Valores faltantes en alguno de \code{xtrain} o \code{ytrain} va a ser droppeado de manera pareada.}

\item{k}{Numero de vecinos cercanos para muestrear desde.}

\item{seed}{Semilla.}
}
\value{
Vector de la misma longitud que \code{xtest} con valores de \code{xtrain}.
}
\description{
Por cada valor en el vector de prediccion \code{xtest},
uno de los cercanos \code{k} valores en el vector de prediccion \code{xtrain}
es elegido aleatoriamente y su valor es observado en \code{ytrain} es retornado.
}
\details{
Predictive Mean Matching
}
\examples{
pmm(xtrain = c(0.2, 0.2, 0.8), xtest = 0.3, ytrain = c(0, 0, 1)) # 0
pmm(xtrain = c(TRUE, FALSE, TRUE), xtest = FALSE, ytrain = c(2, 0, 1)) # 0
pmm(xtrain = c(0.2, 0.8), xtest = 0.3, ytrain = c("A", "B"), k = 2) # "A" or "B"
pmm(xtrain = c("A", "A", "B"), xtest = "A", ytrain = c(2, 2, 4), k = 2) # 2
pmm(xtrain = factor(c("A", "B")), xtest = factor("C"), ytrain = 1:2) # 2
}
\author{
ecoteam2019
}
