% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missRanger.R
\name{revert}
\alias{revert}
\title{Revierte conversion}
\usage{
revert(con, X = con$X)
}
\arguments{
\item{con}{Una lista retornada con \code{convert}.}

\item{X}{Un data frame con algunas columnas a ser convertidas de vuelta de acuerdo con la information almacenada en \code{converted}.}
}
\value{
Un data frame.
}
\description{
Revierte la conversion hecha por \code{convert}.
}
\author{
ecoteam2019
}
