% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missRanger.R
\name{typeof2}
\alias{typeof2}
\title{Una version de \code{typeof} internamente usado por \code{missRanger}.}
\usage{
typeof2(object)
}
\arguments{
\item{object}{Algun objeto.}
}
\value{
Un string.
}
\description{
Retorna o el valor "numeric" (double o integer), "factor", "character", "logical", "special" (modo numerico, pero cualquier double o integer) o "" (cualquier otro).
\code{missRanger} requiere esta informacion para tratar con respuestas tipo no originalmente soportado por \code{ranger}.
}
\author{
ecoteam2019
}
